👌 IMPROVE: Scaffolding + dynamic ABI
